+++ X10D/src/Math/Int32Extensions.cs
@@ -10,6 +10,21 @@ namespace X10D.Math;
 public static class Int32Extensions
 {
     /// <summary>
+    ///     Returns the number of digits in the current 32-bit signed integer.
+    /// </summary>
+    /// <param name="value">The value whose digit count to compute.</param>
+    /// <returns>The number of digits in <paramref name="value" />.</returns>
+    public static int CountDigits(this int value)
+    {
+        if (value == 0)
+        {
+            return 1;
+        }
+
+        return ((long)value).CountDigits();
+    }
+
+    /// <summary>
     ///     Computes the digital root of this 32-bit integer.
     /// </summary>
     /// <param name="value">The value whose digital root to compute.</param>